requirements.txt
lxml
requests
selenium
Pillow